+++ db/schema.rb
@@ -203,11 +203,11 @@ ActiveRecord::Schema.define(version: 20190307151207) do
     t.integer "class_year"
     t.datetime "tos_accepted"
     t.bigint "college_id"
-    t.integer "student_sid"
+    t.integer "student_id"
     t.index ["college_id"], name: "index_users_on_college_id"
     t.index ["email"], name: "index_users_on_email", unique: true
     t.index ["reset_password_token"], name: "index_users_on_reset_password_token", unique: true
-    t.index ["student_sid"], name: "index_users_on_student_sid", unique: true
+    t.index ["student_id"], name: "index_users_on_student_id", unique: true
     t.index ["username"], name: "index_users_on_username", unique: true
   end