+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 All notable changes to this project will be documented in this file.
 
 ## Unreleased
+### Added
+* Add SID to users ([#669](https://gitlab.com/yale-sdmp/vesta/issues/669)).
+
 ### Changed
 * Hide the login link in the nav when a user is not in a specific college ([#909](https://gitlab.com/yale-sdmp/vesta/issues/909)).
 * Hide the draws dropdown if there are no draws ([#884](https://gitlab.com/yale-sdmp/vesta/issues/884)).